Modularize RosbridgeWebSocket
