+++ rosbridge_server/scripts/rosbridge_websocket.py
@@ -34,81 +34,14 @@
 import rospy
 import sys
 
-from rosauth.srv import Authentication
-
-from signal import signal, SIGINT, SIG_DFL
-from functools import partial
 from socket import error
 
 from tornado.ioloop import IOLoop
 from tornado.ioloop import PeriodicCallback
 from tornado.web import Application
-from tornado.websocket import WebSocketHandler
 
-from rosbridge_library.rosbridge_protocol import RosbridgeProtocol
-from rosbridge_library.util import json
+from rosbridge_server import RosbridgeWebSocket
 
-# Global ID seed for clients
-client_id_seed = 0
-clients_connected = 0
-# if authentication should be used
-authenticate = False
-
-class RosbridgeWebSocket(WebSocketHandler):
-
-    def open(self):
-        global client_id_seed, clients_connected, authenticate
-        try:
-            self.protocol = RosbridgeProtocol(client_id_seed)
-            self.protocol.outgoing = self.send_message
-            self.set_nodelay(True)
-            self.authenticated = False
-            client_id_seed = client_id_seed + 1
-            clients_connected = clients_connected + 1
-        except Exception as exc:
-            rospy.logerr("Unable to accept incoming connection.  Reason: %s", str(exc))
-        rospy.loginfo("Client connected.  %d clients total.", clients_connected)
-        if authenticate:
-            rospy.loginfo("Awaiting proper authentication...")
-
-    def on_message(self, message):
-        global authenticate
-        # check if we need to authenticate
-        if authenticate and not self.authenticated:
-            try:
-                msg = json.loads(message)
-                if msg['op'] == 'auth':
-                    # check the authorization information
-                    auth_srv = rospy.ServiceProxy('authenticate', Authentication)
-                    resp = auth_srv(msg['mac'], msg['client'], msg['dest'],
-                                                  msg['rand'], rospy.Time(msg['t']), msg['level'],
-                                                  rospy.Time(msg['end']))
-                    self.authenticated = resp.authenticated
-                    if self.authenticated:
-                        rospy.loginfo("Client %d has authenticated.", self.protocol.client_id)
-                        return
-                # if we are here, no valid authentication was given
-                rospy.logwarn("Client %d did not authenticate. Closing connection.",
-                              self.protocol.client_id)
-                self.close()
-            except:
-                # proper error will be handled in the protocol class
-                self.protocol.incoming(message)
-        else:
-            # no authentication required
-            self.protocol.incoming(message)
-
-    def on_close(self):
-        global clients_connected
-        clients_connected = clients_connected - 1
-        self.protocol.finish()
-        rospy.loginfo("Client disconnected. %d clients total.", clients_connected)
-
-    def send_message(self, message):
-        IOLoop.instance().add_callback(partial(self.write_message, message))
-
-    def check_origin(self, origin):
-        return True
 
 def shutdown_hook():
     IOLoop.instance().stop()
@@ -121,7 +54,7 @@ if __name__ == "__main__":
     certfile = rospy.get_param('~certfile', None)
     keyfile = rospy.get_param('~keyfile', None)
     # if authentication should be used
-    authenticate = rospy.get_param('~authenticate', False)
+    RosbridgeWebSocket.authenticate = rospy.get_param('~authenticate', False)
     port = rospy.get_param('~port', 9090)
     address = rospy.get_param('~address', "")
