+++ rosbridge_server/src/rosbridge_server/__init__.py
@@ -1,0 +1,2 @@
+from websocket_handler import RosbridgeWebSocket
+
